Rename panel_controller to parameterized and move int to interface module
