+++ vivarium/environment/components/entities/interface.py
@@ -8,7 +8,7 @@ from bokeh.models import BooleanFilter, CDSView
 
 import numpy as np
 
-from vivarium.controllers.panel_controller import ParameterizedData
+from vivarium.interface.parameterized import ParameterizedData
 from vivarium.environment.components.interface import Interface, Renderer
