+++ vivarium/environment/components/entities/walls/interface.py
@@ -2,7 +2,7 @@ import param
 from panel.layout import Column
 
 from vivarium.environment.components.interface import Interface, Renderer
-from vivarium.controllers.panel_controller import ParameterizedData
+from vivarium.interface.parameterized import ParameterizedData
 
 class WallRenderer(Renderer):
